requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
et-xmlfile==1.1.0
Flask==2.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.1.1
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.2
psycopg2-binary==2.9.9
PyMySQL==1.1.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
qrcode==7.4.2
simple-websocket==1.0.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==2.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2
